+++ src/main/java/br/com/alopes/deliveryfood/infra/repository/KitchenRepositoryImpl.java
@@ -2,6 +2,7 @@ package br.com.alopes.deliveryfood.infra.repository;
 
 import br.com.alopes.deliveryfood.domain.model.Kitchen;
 import br.com.alopes.deliveryfood.domain.repository.KitchenRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,8 +34,12 @@ public class KitchenRepositoryImpl implements KitchenRepository {
 
     @Transactional
     @Override
-    public void delete(Kitchen kitchen) {
-        kitchen = findById(kitchen.getId());
+    public void delete(Long id) {
+        Kitchen kitchen = findById(id);
+
+        if (kitchen == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(kitchen);
     }
 }
